PackMule: add limited RR policy
This adds a new PackMule rule syntax: `RR(<num>/<player>)`, which means:
"when everyone has received n items in round-robin, then all subsequent
ones go to a designated player".

For example, if Alice, Bob and Calvin are in a group, Alice is loot master,
and has the following PackMule rule:

* `*Hakkari Bijou` to `RR(2/Calvin)`

The first 6 bijous to drop will be distributed in 2 rounds (exactly like with
the current `RR` placeholder), so each member has 2 bijous. Then, all the
subsequent drops go to Calvin.